requirements.txt
requests>=2.31.0
pathlib>=1.0.1
langchain-core>=0.1.27
groq>=0.4.2
python-dotenv>=1.0.1
json-repair>=0.8.1
google-auth>=2.28.1
google-auth-oauthlib>=1.2.0
pandas>=2.2.1
supabase>=2.4.0
sseclient-py>=1.8.0
